requirements.txt
Flask>=3.0.3,<4.0
Pillow>=10.4.0,<11.0
numpy>=1.19,<1.26